Переносит валидацию в отдельный файл
