+++ js/main.js
@@ -3,6 +3,6 @@ import { MAX_PHOTOCARD_AMOUNT } from './data.js';
 import './create-comment.js';
 import './previews.js';
 import './image-upload-form.js';
-
+import './validation.js';
 
 setPhotocard(MAX_PHOTOCARD_AMOUNT);
